refactor: rename component -> components
