+++ example/vite-demo/colorgen/components/Loading.tsx
