+++ example/vite-demo/colorgen/components/ShowInfer.tsx
